FIX: Overflow when generating forests for very large maps
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@9715 8aca7d54-2c30-db11-9de9-000461428c89

+++ bauer/tree_builder.cc
@@ -301,12 +301,12 @@ void tree_builder_t::distribute_trees(int dichte, sint16 xtop, sint16 ytop, sint
 	settings_t const& s             = welt->get_settings();
 	sint32     const  x             = welt->get_size().x;
 	sint32     const  y             = welt->get_size().y;
-	unsigned   const t_forest_size  = (unsigned)pow(((double)x * (double)y), 0.25) * s.get_forest_base_size() / 11 + (x + y) / (2 * s.get_forest_map_size_divisor());
-	uint8      const c_forest_count = (unsigned)pow(((double)x * (double)y), 0.5)  / s.get_forest_count_divisor();
+	unsigned   const t_forest_size  = (uint32)pow(((double)x * (double)y), 0.25) * s.get_forest_base_size() / 11 + (x + y) / (2 * s.get_forest_map_size_divisor());
+	uint32     const c_forest_count = (uint32)pow(((double)x * (double)y), 0.5)  / s.get_forest_count_divisor();
 
 	DBG_MESSAGE("tree_builder_t::distribute_trees", "Creating %i forests", c_forest_count);
 
-	for (uint8 c1 = 0 ; c1 < c_forest_count ; c1++) {
+	for (uint32 c1 = 0; c1 < c_forest_count ; c1++) {
 		// to have same execution order for simrand
 		koord const start = koord::koord_random(x, y);
 		koord const size  = koord(t_forest_size,t_forest_size) + koord::koord_random(t_forest_size, t_forest_size);
